size_t to uint64_t

+++ src/physlr-molecules.cc
@@ -43,7 +43,7 @@ struct vertexProperties
 {
 	std::string name = "";
 	int weight = 0;
-	size_t indexOriginal = 0;
+	uint64_t indexOriginal = 0;
 };
 
 struct edgeProperties
@@ -69,22 +69,22 @@ using graph_t = boost::subgraph<boost::adjacency_list<
     boost::property<
         boost::edge_index_t,
         int,
-        boost::property<edgeComponent_t, std::size_t, edgeProperties>>>>;
+        boost::property<edgeComponent_t, std::uint64_t, edgeProperties>>>>;
 using vertex_t = graph_t::vertex_descriptor;
 using edge_t = graph_t::edge_descriptor;
 using barcodeToIndex_t = std::unordered_map<std::string, vertex_t>;
 using indexToBarcode_t = std::unordered_map<vertex_t, std::string>;
 using vertexSet_t = std::unordered_set<vertex_t>;
 using componentToVertexSet_t = std::vector<vertexSet_t>;
-using vertexToComponent_t = std::unordered_map<vertex_t, size_t>;
+using vertexToComponent_t = std::unordered_map<vertex_t, uint64_t>;
 using vecVertexToComponent_t = std::vector<vertexToComponent_t>;
 using vertexToIndex_t =
-    std::unordered_map<vertex_t, size_t>; // wanna improve this? checkout boost::bimap
+    std::unordered_map<vertex_t, uint64_t>; // wanna improve this? checkout boost::bimap
 using indexToVertex_t =
-    std::unordered_map<size_t, vertex_t>; // wanna improve this? checkout boost::bimap
+    std::unordered_map<uint64_t, vertex_t>; // wanna improve this? checkout boost::bimap
 using adjacencyMatrix_t = std::vector<std::vector<uint_fast32_t>>;
 using adjacencyVector_t = std::vector<uint_fast32_t>;
-using Clique_type = std::unordered_map<vertex_t, size_t>;
+using Clique_type = std::unordered_map<vertex_t, uint64_t>;
 
 static void
 printVersion()
@@ -227,9 +227,9 @@ componentsToNewGraph(
 #if _OPENMP
 	double sTime = omp_get_wtime();
 #endif
-	for (size_t i = 0; i < vecVertexToComponent.size(); ++i) {
+	for (uint64_t i = 0; i < vecVertexToComponent.size(); ++i) {
 
-		size_t maxVal = 0;
+		uint64_t maxVal = 0;
 		if (!vecVertexToComponent[i].empty()) {
 			maxVal =
 			    std::max_element(
@@ -240,7 +240,7 @@ componentsToNewGraph(
 			        ->second;
 		}
 
-		for (size_t j = 0; j < maxVal + 1; ++j) {
+		for (uint64_t j = 0; j < maxVal + 1; ++j) {
 			auto u = boost::add_vertex(molSepG);
 			molSepG[u].name = inG[i].name + "_" + std::to_string(j);
 			molSepG[u].weight = inG[i].weight;
@@ -295,7 +295,7 @@ biconnectedComponents(graph_t& subgraph, vertexToComponent_t& vertexToComponent)
 	componentToVertexSet_t componentToVertexSet;
 
 	for (boost::tie(ei, ei_end) = boost::edges(subgraph); ei != ei_end; ++ei) {
-		size_t componentNum = component[*ei];
+		uint64_t componentNum = component[*ei];
 		if (componentNum + 1 > componentToVertexSet.size()) {
 			componentToVertexSet.resize(componentNum + 1);
 		}
@@ -311,7 +311,7 @@ biconnectedComponents(graph_t& subgraph, vertexToComponent_t& vertexToComponent)
 		}
 	}
 
-	size_t moleculeNum = 0;
+	uint64_t moleculeNum = 0;
 
 	// Remove components with size less than 1
 	for (auto&& vertexSet : componentToVertexSet) {
@@ -348,9 +348,9 @@ make_subgraph(Graph& g, Graph& subgraph, edgeSet& edge_set, vertexIter vBegin, v
 		for (boost::tie(vIter2, vend2) = vertices(subgraph); vIter2 != vend2; ++vIter2) {
 			if (vIter1 != vIter2) {
 				//		        tsl::robin_map<
-				//		            std::pair<std::size_t,size_t>, int,
-				//		                boost::hash<std::pair<size_t,size_t>>>::const_iterator
-				auto got = edge_set.find(std::pair<std::size_t, size_t>(
+				//		            std::pair<std::uint64_t,uint64_t>, int,
+				//		                boost::hash<std::pair<uint64_t,uint64_t>>>::const_iterator
+				auto got = edge_set.find(std::pair<std::uint64_t, uint64_t>(
 				    subgraph[*vIter1].indexOriginal, subgraph[*vIter2].indexOriginal));
 
 				if (got != edge_set.end()) {
@@ -369,7 +369,7 @@ main(int argc, char* argv[])
 	int optindex = 0;
 	static int help = 0;
 	std::string separationStrategy = "bc";
-	size_t threads = 1;
+	uint64_t threads = 1;
 	bool verbose = false;
 	bool failed = false;
 	static const struct option longopts[] = {
@@ -434,7 +434,7 @@ main(int argc, char* argv[])
 #endif
 
 	// // auxillary dataset: set of edges for faster lookup
-	tsl::robin_map<std::pair<std::size_t, size_t>, int, boost::hash<std::pair<size_t, size_t>>>
+	tsl::robin_map<std::pair<std::uint64_t, uint64_t>, int, boost::hash<std::pair<uint64_t, uint64_t>>>
 	    edge_set;
 	edge_set.reserve(num_edges(g));
 
@@ -443,7 +443,7 @@ main(int argc, char* argv[])
 		auto& weight = g[*edgeIt].weight;
 		auto& node1 = g[boost::source(*edgeIt, g)].indexOriginal;
 		auto& node2 = g[boost::target(*edgeIt, g)].indexOriginal;
-		edge_set[std::pair<size_t, size_t>(node1, node2)] = weight;
+		edge_set[std::pair<uint64_t, uint64_t>(node1, node2)] = weight;
 	}
 
 	auto vertexItRange = vertices(g);
